web app changes

+++ lib/services/functions.dart
@@ -135,7 +135,7 @@ class ServiceClass {
     return response;
   }
 
-  Future<String> decayTokens(String customerAddress,int points,Web3Client ethClient) async{
+  Future<String> decayTokens(String customerAddress,BigInt points,Web3Client ethClient) async{
     var response=await _callFunction('decayTokens',[EthereumAddress.fromHex(customerAddress),
       points],ethClient,owner_private_key);
     print('Function decayTokens Called Successfully');
